refactor: render 함수 모듈화
